Remove some warnings

+++ src/Crypto/Hash/ExcelHash.cpp
@@ -43,7 +43,7 @@ void Crypto::Hash::ExcelHash::Calc(const UInt8 *buff, UOSInt buffSize)
 		this->currVal = (UInt16)((this->currVal ^ buff[buffSize]) << 1);
 		if (this->currVal & 0x8000)
 		{
-			this->currVal = (this->currVal & 0x7fff) | 1;
+			this->currVal = (UInt16)((this->currVal & 0x7fff) | 1);
 		}
 	}
 }
